restaurant_advisor/requirements.txt
langchain
langchain-community
langchain-core
langchain-mongodb
langchain-google-genai
langsmith
langgraph
sentence-transformers
neo4j
pymongo
python-dotenv
google-generativeai
pypdf
rich
typer
python-jose
passlib
bcrypt
